requirements-api.txt
# API/Web Framework Requirements
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
pyyaml==6.0.1
